Implemented runtime environment for Python engines
This is supposed to be run by services, but you can also
experiment with Python engines in interactive mode, e.g.:

$ python -c "from onlinelab.service.engine.python import
    PythonEngine; PythonEngine().run(interactive=True)"
port=39660, pid=19773

(make sure onlinelab is on Python's PATH)

On stdout we got port number (on which the engine is listening
for requests) and process ID. To communicate with this engine
you have to start another terminal and run python (or ipython):

In [1]: from xmlrpclib import ServerProxy

In [2]: s = ServerProxy("http://localhost:39660")

In [3]: s.evaluate("2 + 2")
Out[3]:
{'err': '',
 'interrupted': False,
 'out': '4\n',
 'source': '2 + 2',
 'traceback': False}
In [4]: s.evaluate("2 +")
Out[4]:
{'err': '',
 'interrupted': False,
 'out': '',
 'source': '2 +',
 'traceback': '  File "<online-lab>", line 1\n    2 +\n      ^\n
 SyntaxError: unexpected EOF while parsing\n'}

In [5]: s.evaluate("x")
Out[5]:
{'err': '',
 'interrupted': False,
 'out': '',
 'source': 'x',
 'traceback': 'Traceback (most recent call last):\n  File "<online-lab>",
 line 1, in <module>\nNameError: name \'x\' is not defined\n'}

In [6]: s.evaluate("while 1: pass")

This hangs the interpreter, because SeverProxy is waiting for
a response from the engine. To interrupt evaluation, send INT
signal (#2) to engine process, e.g. in another terminal issue
'kill -s INT 19773' and we obtain the following response:

Out[6]:
{'err': '',
'interrupted': True,
'out': '',
'source': 'while 1: pass',
'traceback': False}

You can also run engines in non-interactive mode (default), but
then you won't be able to terminate engine process with Ctrl+C.
To achieve this send TERM signal to the engine you want to stop.