Added support for single user mode (lab)
Add --no-auth to 'core start' command line. A new user 'lab'
will be created automatically without password. You will be
able to logout and login with 'lab' user anytime. To disable
'lab' user, run core without --no-auth. The user will be kept
in the database but it won't be possible to log in with or
without password. Thus, 'lab' account will be special from
now on.