ifemhub: don't raise an exception if kill() fails
The engine is killed whenever the Console() class goes out of scope. Which
is the case when some exception happens (like if the server engine crashes). In
such situations, also the engine kill fails (raising some exception). So we
catch it, and just print "Failed", so that the user knows that the engine kill
failed, and the original exception (from the evaluation) still gets propagated
and shown to the user.